# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Evatrack API",
        "description": "API Documentation for Evatrack Evacuation Management System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/analytics/dashboard": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get dashboard analytics data",
                "operationId": "8647472c701e5a2ced5dc4af8550284c",
                "parameters": [
                    {
                        "name": "event_id",
                        "in": "query",
                        "description": "Filter by event ID or \"all\"",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "all"
                        }
                    },
                    {
                        "name": "center_id",
                        "in": "query",
                        "description": "Filter by center ID or \"all\"",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "all"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Filter from date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Filter to date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/analytics/events-list": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get list of disaster events for dropdowns",
                "operationId": "7095643c401daf28d138c930f1c4e60c",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/analytics/event/{eventId}": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get analytics for a specific event (legacy)",
                "operationId": "43f4e52c8a3ef108b69eb9fff4d8985e",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/analytics/event/{eventId}/center/{centerId}": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get analytics for a specific event at a center (legacy)",
                "operationId": "bd1dab975ee0911bcca19929ac3e50ff",
                "parameters": [
                    {
                        "name": "eventId",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "centerId",
                        "in": "path",
                        "description": "Center ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Event or Center not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/user": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get current logged in user",
                "operationId": "b01c1e5cc61cbd59bd13ef6ddc2d9e6d",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log in user",
                "operationId": "4b16aff1c94e328a8af8ef6811b43c2e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "password"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid user ID or password"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log out user",
                "operationId": "dbf312e3830d639842de6d8f314c3f57",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}